% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{compile_model}
\alias{compile_model}
\title{Format output of n break model}
\usage{
compile_model(x, digits = 3)
}
\arguments{
\item{x}{the `model` class to format}

\item{digits}{number of digits displayed in console. Default value is 3}
}
\value{
`x` The input list of `model` class contains the following tables:
\itemize{
\item{`date_tab` A data frame storing the break date estimated by the model,
and their corresponding confidence intervals}
\item{`RS_tab` A data frame storing the estimated coefficients which allowed to
change across regimes with corrected standard errors}
\item{`FS_tab` A data frame storing the estimated coefficients which is constant
across regimes with corrected standard errors}}
}
\description{
`compile_model()` compiles the information of `model` class object `x` into 3
main tables:
\itemize{
\item {`date_tab` Table for estimated break date in the model
with 90% and 95% confidence intervals based on `robust`,`hetomega`, `hetq` options
for errors
and `prewhit` option}
\item {`RS_tab` Table for estimated coefficients for `z` regressors
with corrected standard errors based on `robust`,`hetdat`,`hetvar`
options for errors and `prewhit` option}
\item {`FS_tab` Table for estimated coefficients for `x` regressors
with corrected standard errors based on `robust`,`hetdat`,`hetvar` options for errors and `prewhit` option
}}
}
\note{
\itemize{
\item{If `x` returns 0 number of estimated break,
the function will return `NULL` value instead of the list in `Value`.}
\item{If `x` is a pure structural break, the `FS_tab` will return `NULL` in
`Value`}
}
}
