% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{print.mb}
\alias{print.mb}
\title{print function for structural break model (basic block)}
\usage{
\method{print}{mb}(s, digits = 2, nsmall = 3, ...)
}
\description{
print function for structural break model (basic block)
}
