% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{print.sbm}
\alias{print.sbm}
\title{Custom print method for MyClass}
\usage{
\method{print}{sbm}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class "sbm"}
}
\description{
This method provides a custom printing behavior for objects of class MyClass.
It displays the attribute of the object.
}
