% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{sbm}
\alias{sbm}
\title{Comprehensive structural change estimation and testing}
\usage{
sbm(formula, data, m = 5, eps = 0.15)
}
\arguments{
\item{formula}{an object of class formula: a symbolic description of the pure/partial
structural break model to be fitted. The details of model specification are given 
under 'Details'}

\item{data}{a data frame, or list containing the variables in the model.}

\item{eps1}{trimming level (in percentage) influencing the minimal segment length that is 
equivalent to \code{default} = int(\code{eps1}*T) (T is total sample size).
\itemize{
\item{`eps1=0.05` corresponds to maximum 9 possible breaks}
\item{`eps1=0.10` corresponds to maximum 8 possible breaks}
\item{`eps1=0.15` corresponds to maximum 5 possible break}
\item{`eps1=0.20` corresponds to maximum 3 possible break}
\item{`eps1=0.25` corresponds to maximum 2 possible break}
} If users want to specify their own combinations of number of breaks `m` and minimum segment length `h`, use [estimate()]}

\item{cov.}{There are six parameters users can specify for covariance matrix:
\itemize{
 \item{`prewhit` sets to \code{1} to apply AR(1) prewhitening prior to estimating
the long run covariance matrix. The method used is \emph{Andrews(1991)} automatic bandwidth with AR(1) approximation with quadratic
kernel. Note: Do not set to \code{1} if lagged dependent variables are
included as regressors.}
\item{`robust` sets to \code{1} to allow for heterogeneity and autocorrelation in the residuals, \code{0} otherwise.}
\item{`hetdat` is used for the construction of the F tests. Set to \code{1} if users want to allow different moment 
matrices of the regressors across segments. If \code{hetdat} = \code{0}, the same moment matrices are assumed for each segment
and estimated from the ful sample. It is recommended to set \code{hetdat}=\code{1} if number of regressors \code{x} > \code{0}.}
\item{`hetvar` is used for the construction of the F tests. Set to \code{1}
if users want to allow for the variance of the residuals to be different across segments. If \code{hetvar}=\code{0}, 
the variance of the residuals is assumed constant across segments and constructed from the full sample. 
\code{hetvar}=\code{1} when \code{robust} =\code{1}.}
\item{`hetomega` used in the construction of the confidence intervals for the break
dates. If \code{hetomega}=\code{0}, the long run covariance matrix of zu is
assumed identical across segments (the variance of the errors u if \code{robust}={0}).}
\item{`hetq` used in the construction of the confidence intervals for the break
dates. If \code{hetq}=\code{0}, the moment matrix of the data is assumed identical
across segments.}
}}
}
\value{
`sbm` returns an object of class `sbm`. 
The functions summary is used to obtain and print a summary. The generic accessor
functions `` extract useful features of the value returned by `sbm`. An object of
class `sbm` is a list containing at least the following components:
}
\description{
`sbm()` is used to estimate linear models with structural breaks, including both pure and
partial model with pre-specified number of breaks given a pre-specified trimming level. It
can be used to carry out inference for each regime, testing and selecting number of
structural breaks using information criteria.
}
\details{
Structural break models for `sbm` are specified symbolically similar to linear regression 
model specified in `lm`. A typical structural break model has the form `y ~ z | x` where 
response `y` is the (numeric) response vector and `x` and `z` terms is a series of terms which 
are classified in two distinct types. `x` terms have unchanged coefficients across regimes and `z` terms 
have varied coefficients across regimes. A terms specification of the form `z1 + z2 | x1 + x2`  indicates 
all the terms in `z1` with all the terms in `z2` with duplicates removed have coefficients varied 
across regimes, whereas all terms in `x1` with all terms in `x2` with duplicates removed have
coefficients fixed across regimes. If no `x` regressors are specified in the symbolic formula,
the model is pure structural break model, whereas if any `x` regressors are specified, the model is
partial structural break model. A formula has an implied intercept term that varies across regimes. 
If intercept is explicitly removed by using `y ~ z - 1` for pure structural model, or `y ~ z - 1 | x` for partial structural
change model, then the formula has an implied intercept that fixes across regimes. To remove intercept from
the model entirely, use `y ~ z - 1| x - 1`
}
\examples{
model1 <- sbm(inf~inflag+inffut|ygap,data=nkpc)
model2 <- sbm(rate~1,data=real)
}
\seealso{
[dotest()], [doseqtests()], [doorder()], [dosequa()], and [dofix()]
which are functions called by [mdl()].
}
