% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R, R/SupTest.R
\name{spflp1}
\alias{spflp1}
\title{SupF(l+1|l) test}
\usage{
spflp1(bigvec, dt, nseg, y, z, h, q, prewhit, robust, x, p, hetdat, hetvar)

spflp1(bigvec, dt, nseg, y, z, h, q, prewhit, robust, x, p, hetdat, hetvar)
}
\arguments{
\item{bigvec}{associated SSR of estimated break date under H0}

\item{dt}{estimated date under H0}

\item{nseg}{number of segment under H1}

\item{y}{matrix of dependent variable}

\item{z}{independent variables with coefficients are allowed to change across
regimes}

\item{h}{minimum segment length}

\item{q}{number of \code{z} regressors}

\item{prewhit, robust, hetdat, hetvar}{options on residuals/errors}

\item{x}{independent variables with constant coefficients across regimes}

\item{p}{number of \code{x} regressors}
}
\value{
A list that contains the following:
\itemize{
\item {maxf}{Maximum value of test}
\item{newd}{Additional date in alternative hypothesis }
}

A list that contains the following:
\itemize{
\item {maxf}{Maximum value of test}
\item{newd}{Additional date in alternative hypothesis }
}
}
\description{
`spflp1` computes the test statistics of supF(l+1|l) test with null hypothesis
is l=`nseg-1` and alternative hypothesis is l+1.
The l breaks under the null hypothesis are taken from the global minimization.

Function computes the test statistics of supF(l+1|l) test with null hypothesis
is l=\code{nseg}-1 and alternative hypothesis is l+1.
The l breaks under the null hypothesis are taken from the global minimization.
}
